requirements.txt
tabulate==0.9.0
coinbase-advanced-py==1.5.0
python-dotenv==0.21.0